requirements.txt
# Main packages for API interactions and ML functionality
huggingface_hub>=0.19.0
requests>=2.31.0
python-dotenv>=1.0.0

# Google AI/LLM integration
google-generativeai>=0.3.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
